Reorganize to use conf for module loading
Ill keep the technical modules configured in code in the app module but
move to have the feature modules loaded via conf.
This also regoranizes the database module a little bit to use less conf
injection. I'd like to get rid of all of it but I have too many
properties right now for the DI engine's provide fun as far as i can
tell when i try to do it purely in code.